Use existing Gender model for patient. Changed task status from json to simple enum. Moved patient models to base image.
