+++ scanhub-ui/src/components/TaskCreateModal.tsx
@@ -37,7 +37,7 @@ export default function TaskCreateModal(props: ModalPropsCreate) {
     name: '',
     description: '',
     type: TaskType.ProcessingTask,
-    status: {},
+    status: 'PENDING',
     args: {},
     artifacts: {},
     destinations: {},
