+++ scanhub-ui/src/generated-client/exam/api.ts
@@ -143,10 +143,10 @@ export interface BaseTask {
     'destinations': { [key: string]: string; };
     /**
      * 
-     * @type {{ [key: string]: string; }}
+     * @type {TaskStatus}
      * @memberof BaseTask
      */
-    'status': { [key: string]: string; };
+    'status': TaskStatus;
     /**
      * 
      * @type {boolean}
@@ -383,10 +383,10 @@ export interface TaskOut {
     'destinations': { [key: string]: string; };
     /**
      * 
-     * @type {{ [key: string]: string; }}
+     * @type {TaskStatus}
      * @memberof TaskOut
      */
-    'status': { [key: string]: string; };
+    'status': TaskStatus;
     /**
      * 
      * @type {boolean}
@@ -412,6 +412,23 @@ export interface TaskOut {
      */
     'datetime_created': string;
 }
+
+
+/**
+ * Task status enum.
+ * @export
+ * @enum {string}
+ */
+
+export const TaskStatus = {
+    Pending: 'PENDING',
+    InProgress: 'IN_PROGRESS',
+    Completed: 'COMPLETED',
+    Failed: 'FAILED',
+    Error: 'ERROR'
+} as const;
+
+export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];
 
 
 /**
